Various changes.

+++ win/win32/sdl/sdlmisc.h
@@ -7,7 +7,7 @@
 
 extern unsigned char* image_load_from_resource(HINSTANCE hInstance, int resource_id, int* x_ptr, int* y_ptr, size_t* n_ptr);
 extern GLuint gl2_image_load_from_resource(HINSTANCE hInstance, int resource_id, int* x_ptr, int* y_ptr, size_t* n_ptr);
-extern GLuint gl2_image_load(const char* filename, int* x_ptr, int* y_ptr, int* n_ptr);
+extern GLuint gl2_image_load(const char* filename, int* x_ptr, int* y_ptr, size_t* n_ptr);
 extern SDL_Texture* sdl_texture_from_resource(SDL_Renderer* renderer, HINSTANCE hInstance, int resource_id);
 extern SDL_Surface* sdl_surface_from_resource(HINSTANCE hInstance, int resource_id);
